requirements.txt
appdirs==1.4.3
click==6.7
Flask==0.12.1
future==0.16.0
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.9.6
line-bot-sdk==1.1.0
MarkupSafe==1.0
packaging==16.8
pyparsing==2.2.0
requests==2.13.0
six==1.10.0
Werkzeug==0.12.1
